refactor(pdf): extract region caption analysis
